requirements.txt
# Requirements python==3.8

# --------- hydra --------- #
hydra-core==1.1.0.dev5
hydra-colorlog==1.0.0
hydra-optuna-sweeper==0.9.0rc2
# --------- loggers --------- #
wandb>=0.10.20
# --------- segmentation --------- #
tifffile
numpy
scipy
scikit-image
keras
networkx
SimpleITK
tensorflow
tensorflow-gpu
imagecodecs
pyyaml
# --------- others --------- #
# scikit-fmm==2021.2.2
git+https://github.com/watarungurunnn/scikit-image
